get login response and update assets API
