+++ background/assets.ts
@@ -76,6 +76,17 @@ export type FungibleAsset = Asset & {
 export type FiatCurrency = FungibleAsset
 
 /**
+ * Represents an asset that is stored in an off-chain acount.
+ * This asset may be a fiat or cryptocurrency.
+ */
+export type OffChainAsset = {
+  amount: number,
+  currencySymbol: string,
+  label: string,
+  accountId: string
+}
+
+/**
  * Any asset that exists on a particular network; see {@link NetworkSpecific)
  * for information on network-specific objects.
  */
