+++ ui/services/OffChainService.ts
@@ -1,9 +1,9 @@
-import { OffChainAccountCredentials, OffChainProvider } from "@tallyho/tally-background/accounts";
-import { OffChainAsset } from "@tallyho/tally-background/redux-slices/assets";
+import { OffChainAccount, OffChainAccountCredentials, OffChainProvider } from "@tallyho/tally-background/accounts";
+import { OffChainAsset } from "@tallyho/tally-background/assets";
 
 export class OffChainService {
 
-    static login({provider, credentials }: {provider: OffChainProvider,credentials: OffChainAccountCredentials}): Promise<{assets: OffChainAsset[]}> {
+    static login({provider, credentials }: {provider: OffChainProvider,credentials: OffChainAccountCredentials}): Promise<OffChainAccount> {
         const apiResponsePromise = fetch(`https://mocki.io/v1/d14997dd-ea1d-496a-b3ca-f6576201a9f3`, {
         // const apiResponsePromise = fetch(`${provider}/api/v1/login`, {
             method: 'GET',
@@ -25,13 +25,15 @@ export class OffChainService {
     }
 
     static assets({ userId = ""}): Promise<{assets: OffChainAsset[]}> {
-        // const apiResponsePromise = fetch(`https://mocki.io/v1/13770a8a-231a-46fd-9acd-f337b42e0436`, {
+        const token = localStorage.getItem('token');
+        // const apiResponsePromise = fetch(`https://mocki.io/v1/6f412ee4-2875-4764-bbdc-eefada21ec2a`, {
         const apiResponsePromise = fetch(`http://192.168.1.120:8000/api/v1/balances?user_id=${userId}`, {
             method: 'GET',
             headers: {
             'Accept': 'application/json',
             'Content-Type': 'application/json',
-            'Referrer': document.location.origin
+            'Referrer': document.location.origin,
+            'Authorization': `JWT ${token}`,
             }
         })
         .then(response => {
